+++ R/dada_phyloseq.R
@@ -111,9 +111,11 @@ clean_pq <- function(physeq,
   }
 
   if (sum(grepl("^0", sample_names(physeq)) > 0) && !silent) {
-    message("At least one sample name start with a zero.
+    message(
+      "At least one sample name start with a zero.
     That can be a problem for some phyloseq functions such as
-    plot_bar and psmelt.")
+    plot_bar and psmelt."
+    )
   }
 
   if (force_taxa_as_columns && force_taxa_as_rows) {
@@ -121,16 +123,20 @@ clean_pq <- function(physeq,
   }
 
   if (force_taxa_as_columns && taxa_are_rows(physeq)) {
-    otu_table(physeq) <- otu_table(t(as.matrix(unclass(physeq@otu_table))),
-      taxa_are_rows = FALSE
-    )
+    otu_table(physeq) <-
+      otu_table(t(as.matrix(unclass(
+        physeq@otu_table
+      ))),
+      taxa_are_rows = FALSE)
     message("Taxa are now in columns.")
   }
 
   if (force_taxa_as_rows && !taxa_are_rows(physeq)) {
-    otu_table(physeq) <- otu_table(t(as.matrix(unclass(physeq@otu_table))),
-      taxa_are_rows = TRUE
-    )
+    otu_table(physeq) <-
+      otu_table(t(as.matrix(unclass(
+        physeq@otu_table
+      ))),
+      taxa_are_rows = TRUE)
     message("Taxa are now in rows.")
   }
 
@@ -152,18 +158,29 @@ clean_pq <- function(physeq,
   }
 
   if (verbose) {
-    message(paste(
-      "Cleaning suppress", ntaxa(physeq) - ntaxa(new_physeq), "taxa (",
-      paste(taxa_names(physeq)[taxa_sums(physeq) == 0], collapse = " / "), ") and",
-      nsamples(physeq) - nsamples(new_physeq),
-      "sample(s) (", paste(sample_names(physeq)[sample_sums(physeq) == 0], collapse = " / "), ")."
-    ))
+    message(
+      paste(
+        "Cleaning suppress",
+        ntaxa(physeq) - ntaxa(new_physeq),
+        "taxa (",
+        paste(taxa_names(physeq)[taxa_sums(physeq) == 0], collapse = " / "),
+        ") and",
+        nsamples(physeq) - nsamples(new_physeq),
+        "sample(s) (",
+        paste(sample_names(physeq)[sample_sums(physeq) == 0], collapse = " / "),
+        ")."
+      )
+    )
   } else if (!silent) {
-    message(paste(
-      "Cleaning suppress", ntaxa(physeq) - ntaxa(new_physeq), "taxa and",
-      nsamples(physeq) - nsamples(new_physeq),
-      "samples."
-    ))
+    message(
+      paste(
+        "Cleaning suppress",
+        ntaxa(physeq) - ntaxa(new_physeq),
+        "taxa and",
+        nsamples(physeq) - nsamples(new_physeq),
+        "samples."
+      )
+    )
   }
 
   verify_pq(new_physeq)
@@ -205,12 +222,11 @@ clean_pq <- function(physeq,
 #'   each object.
 #' @export
 
-track_wkflow <- function(
-    list_of_objects,
-    obj_names = NULL,
-    clean_pq = FALSE,
-    taxonomy_rank = NULL,
-    ...) {
+track_wkflow <- function(list_of_objects,
+                         obj_names = NULL,
+                         clean_pq = FALSE,
+                         taxonomy_rank = NULL,
+                         ...) {
   message("Compute the number of sequences")
   if (!is.null(obj_names)) {
     names(list_of_objects) <- obj_names
@@ -232,19 +248,17 @@ track_wkflow <- function(
       } else if (inherits(object, "matrix")) {
         sum(object, na.rm = TRUE)
       } else if (is.character(object[1]) &&
-        length(object[1]) == 1 &&
-        file.exists(object[1])) {
+                 length(object[1]) == 1 &&
+                 file.exists(object[1])) {
         if (summary(file(object[[1]]))$class == "gzfile") {
           pbapply::pbsapply(object, function(x) {
             as.numeric(system(paste("zcat ", x, " | grep -c '^+$'", sep = ""),
-              intern = TRUE
-            ))
+                              intern = TRUE))
           })
         } else if (grepl("\\.fastq$", object[1])) {
           pbapply::pbsapply(object, function(x) {
             as.numeric(system(paste("cat ", x, " | grep -c '^+$'", sep = ""),
-              intern = TRUE
-            ))
+                              intern = TRUE))
           })
         } else {
           stop("Files must be either gzfile or .fastq")
@@ -275,7 +289,7 @@ track_wkflow <- function(
       } else if (inherits(object[[1]], "dada")) {
         dim(suppressMessages(dada2::makeSequenceTable(object)))[2]
       } else if (is.data.frame(object[[1]]) &&
-        all(c("sequence", "abundance") %in% colnames(object[[1]]))) {
+                 all(c("sequence", "abundance") %in% colnames(object[[1]]))) {
         dim(suppressMessages(dada2::makeSequenceTable(object)))[2]
       } else if (inherits(object, "derep")) {
         length(unique(names(object$uniques)))
@@ -301,15 +315,15 @@ track_wkflow <- function(
       } else if (inherits(object[[1]], "dada")) {
         dim(suppressMessages(dada2::makeSequenceTable(object)))[1]
       } else if (is.data.frame(object[[1]]) &&
-        all(c("sequence", "abundance") %in% colnames(object[[1]]))) {
+                 all(c("sequence", "abundance") %in% colnames(object[[1]]))) {
         dim(suppressMessages(dada2::makeSequenceTable(object)))[1]
       } else if (inherits(object, "derep")) {
         1
       } else if (inherits(object[[1]], "derep")) {
         length(object)
       } else if (is.character(object[1]) &&
-        length(object[1]) == 1 &&
-        file.exists(object[1])) {
+                 length(object[1]) == 1 &&
+                 file.exists(object[1])) {
         length(object)
       } else {
         NA
@@ -323,7 +337,7 @@ track_wkflow <- function(
         message(paste("Start object of class:", class(object), sep = " "))
         if (inherits(object, "phyloseq")) {
           if (taxa_are_rows(object)) {
-            apply(object@tax_table[taxonomy_rank, ], 1, function(x) {
+            apply(object@tax_table[taxonomy_rank,], 1, function(x) {
               length(unique(stats::na.omit(x)))
             })
           } else {
@@ -349,29 +363,34 @@ track_wkflow <- function(
       })
     track <- plyr::rbind.fill.matrix(
       matrix(ncol = length(list_of_objects), unlist(track_nb_seq_per_obj)),
-      matrix(ncol = length(list_of_objects), unlist(track_nb_cluster_per_obj)),
+      matrix(
+        ncol = length(list_of_objects),
+        unlist(track_nb_cluster_per_obj)
+      ),
       matrix(ncol = length(list_of_objects), unlist(track_nb_sam_per_obj)),
-      matrix(ncol = length(list_of_objects), unlist(track_nb_tax_value_per_obj))
+      matrix(
+        ncol = length(list_of_objects),
+        unlist(track_nb_tax_value_per_obj)
+      )
     )
 
-    rownames(track) <- c(
-      "nb_sequences",
-      "nb_clusters",
-      "nb_samples",
-      names_taxonomic_rank[[1]]
-    )
+    rownames(track) <- c("nb_sequences",
+                         "nb_clusters",
+                         "nb_samples",
+                         names_taxonomic_rank[[1]])
   } else {
     track <- plyr::rbind.fill.matrix(
       matrix(ncol = length(list_of_objects), unlist(track_nb_seq_per_obj)),
-      matrix(ncol = length(list_of_objects), unlist(track_nb_cluster_per_obj)),
+      matrix(
+        ncol = length(list_of_objects),
+        unlist(track_nb_cluster_per_obj)
+      ),
       matrix(ncol = length(list_of_objects), unlist(track_nb_sam_per_obj))
     )
 
-    rownames(track) <- c(
-      "nb_sequences",
-      "nb_clusters",
-      "nb_samples"
-    )
+    rownames(track) <- c("nb_sequences",
+                         "nb_clusters",
+                         "nb_samples")
   }
 
 
@@ -511,8 +530,10 @@ asv2otu <- function(physeq = NULL,
   } else if (inherits(seq_names, "character")) {
     dna <- Biostrings::DNAStringSet(seq_names)
   } else {
-    stop("You must set the args physeq (object of class phyloseq) or
-    seq_names (character vector).")
+    stop(
+      "You must set the args physeq (object of class phyloseq) or
+    seq_names (character vector)."
+    )
   }
 
   if (!method %in% c("clusterize", "vsearch")) {
@@ -521,25 +542,24 @@ asv2otu <- function(physeq = NULL,
 
   if (method == "clusterize") {
     ## Find clusters of ASVs to form the new OTUs
-    clusters <- DECIPHER::Clusterize(
-      dna,
-      cutoff = 1 - id,
-      # e.g. `cutoff = 0.03` for a 97% OTU
-      processors = nproc,
-      ...
-    )
+    clusters <- DECIPHER::Clusterize(dna,
+                                     cutoff = 1 - id,
+                                     # e.g. `cutoff = 0.03` for a 97% OTU
+                                     processors = nproc,
+                                     ...)
 
     if (inherits(physeq, "phyloseq")) {
       new_obj <-
         speedyseq::merge_taxa_vec(physeq,
-          clusters$cluster,
-          tax_adjust = tax_adjust
-        )
+                                  clusters$cluster,
+                                  tax_adjust = tax_adjust)
     } else if (inherits(seq_names, "character")) {
       new_obj <- clusters
     } else {
-      stop("You must set the args physeq (object of class phyloseq) or
-    seq_names (character vector).")
+      stop(
+        "You must set the args physeq (object of class phyloseq) or
+    seq_names (character vector)."
+      )
     }
   } else if (method == "vsearch") {
     Biostrings::writeXStringSet(dna, paste0(tempdir(), "/", "temp.fasta"))
@@ -547,7 +567,14 @@ asv2otu <- function(physeq = NULL,
     system2(
       vsearchpath,
       paste0(
-        paste0(" ", vsearch_cluster_method, " ", paste0(tempdir(), "/", "temp.fasta"), " ", vsearch_args),
+        paste0(
+          " ",
+          vsearch_cluster_method,
+          " ",
+          paste0(tempdir(), "/", "temp.fasta"),
+          " ",
+          vsearch_args
+        ),
         " -id ",
         id,
         " --centroids ",
@@ -559,7 +586,8 @@ asv2otu <- function(physeq = NULL,
       stderr = TRUE
     )
 
-    pack_clusts <- utils::read.table(paste0(tempdir(), "/", "temp.uc"), sep = "\t")
+    pack_clusts <-
+      utils::read.table(paste0(tempdir(), "/", "temp.uc"), sep = "\t")
     colnames(pack_clusts) <-
       c(
         "type",
@@ -581,22 +609,26 @@ asv2otu <- function(physeq = NULL,
     if (inherits(physeq, "phyloseq")) {
       new_obj <-
         speedyseq::merge_taxa_vec(physeq,
-          clusters,
-          tax_adjust = tax_adjust
-        )
+                                  clusters,
+                                  tax_adjust = tax_adjust)
     } else if (inherits(seq_names, "character")) {
       new_obj <- pack_clusts
     } else {
-      stop("You must set the args physeq (object of class phyloseq) or seq_names (character vector).")
+      stop(
+        "You must set the args physeq (object of class phyloseq) or seq_names (character vector)."
+      )
     }
 
-    if (file.exists(paste0(tempdir(), "/", "temp.fasta")) && !keep_temporary_files) {
+    if (file.exists(paste0(tempdir(), "/", "temp.fasta")) &&
+        !keep_temporary_files) {
       unlink(paste0(tempdir(), "/", "temp.fasta"))
     }
-    if (file.exists(paste0(tempdir(), "/", "cluster.fasta")) && !keep_temporary_files) {
+    if (file.exists(paste0(tempdir(), "/", "cluster.fasta")) &&
+        !keep_temporary_files) {
       unlink(paste0(tempdir(), "/", "cluster.fasta"))
     }
-    if (file.exists(paste0(tempdir(), "/", "temp.uc")) && !keep_temporary_files) {
+    if (file.exists(paste0(tempdir(), "/", "temp.uc")) &&
+        !keep_temporary_files) {
       unlink(paste0(tempdir(), "/", "temp.uc"))
     }
   }
@@ -638,6 +670,9 @@ asv2otu <- function(physeq = NULL,
 #' }
 #' @return A dataframe with uc results (invisible)
 #' @export
+#' @details
+#' This function is mainly a wrapper of the work of others.
+#'   Please make [vsearch](https://github.com/torognes/vsearch).
 
 vs_search_global <- function(physeq,
                              seq2search = NULL,
@@ -688,7 +723,8 @@ vs_search_global <- function(physeq,
     )
   )
 
-  pack_clusts <- utils::read.table(paste0(tempdir(), "/", "temp.uc"), sep = "\t")
+  pack_clusts <-
+    utils::read.table(paste0(tempdir(), "/", "temp.uc"), sep = "\t")
   colnames(pack_clusts) <- c(
     "type",
     "cluster",
@@ -779,7 +815,8 @@ write_pq <- function(physeq,
                      ...) {
   verify_pq(physeq)
 
-  physeq <- clean_pq(physeq,
+  physeq <- clean_pq(
+    physeq,
     reorder_asv = reorder_asv,
     rename_asv = rename_asv,
     remove_empty_samples = remove_empty_samples,
@@ -793,30 +830,38 @@ write_pq <- function(physeq,
     dir.create(file.path(path), recursive = TRUE)
   }
   if (one_file) {
-    if (!is.null(physeq@refseq) && !is.null(physeq@otu_table) && !is.null(physeq@tax_table)) {
+    if (!is.null(physeq@refseq) &&
+        !is.null(physeq@otu_table) && !is.null(physeq@tax_table)) {
       if (!taxa_are_rows(physeq)) {
-        otu_table(physeq) <- otu_table(t(as.matrix(unclass(physeq@otu_table))),
-          taxa_are_rows = TRUE
-        )
+        otu_table(physeq) <-
+          otu_table(t(as.matrix(unclass(
+            physeq@otu_table
+          ))),
+          taxa_are_rows = TRUE)
       }
-      df_physeq_interm <- cbind(
-        physeq@otu_table,
-        physeq@tax_table,
-        as.vector(physeq@refseq)
-      )
-      colnames(df_physeq_interm) <- c(sample_names(physeq), colnames(physeq@tax_table), "Reference Sequences")
+      df_physeq_interm <- cbind(physeq@otu_table,
+                                physeq@tax_table,
+                                as.vector(physeq@refseq))
+      colnames(df_physeq_interm) <-
+        c(sample_names(physeq),
+          colnames(physeq@tax_table),
+          "Reference Sequences")
 
       df_physeq_interm <- as.data.frame(df_physeq_interm)
 
       if (write_sam_data) {
-        sam_data <- data.frame(t(data.frame(unclass(physeq@sam_data))))
+        sam_data <- data.frame(t(data.frame(unclass(
+          physeq@sam_data
+        ))))
         colnames(sam_data) <- sample_names(physeq)
         if (sam_data_first) {
           df_physeq <- dplyr::full_join(sam_data, df_physeq_interm)
-          rownames(df_physeq) <- c(rownames(sam_data), rownames(df_physeq_interm))
+          rownames(df_physeq) <-
+            c(rownames(sam_data), rownames(df_physeq_interm))
         } else {
           df_physeq <- dplyr::full_join(df_physeq_interm, sam_data)
-          rownames(df_physeq) <- c(rownames(df_physeq_interm), rownames(sam_data))
+          rownames(df_physeq) <-
+            c(rownames(df_physeq_interm), rownames(sam_data))
         }
       } else {
         df_physeq <- df_physeq_interm
@@ -828,29 +873,37 @@ write_pq <- function(physeq,
         sep = sep_csv,
         ...
       )
-    } else if (!is.null(physeq@otu_table) && !is.null(physeq@tax_table)) {
+    } else if (!is.null(physeq@otu_table) &&
+               !is.null(physeq@tax_table)) {
       if (!taxa_are_rows(physeq)) {
-        otu_table(physeq) <- otu_table(t(as.matrix(unclass(physeq@otu_table))),
-          taxa_are_rows = TRUE
-        )
+        otu_table(physeq) <-
+          otu_table(t(as.matrix(unclass(
+            physeq@otu_table
+          ))),
+          taxa_are_rows = TRUE)
       }
-      df_physeq_interm <- cbind(
-        physeq@otu_table,
-        physeq@tax_table,
-      )
-      colnames(df_physeq_interm) <- c(sample_names(physeq), colnames(physeq@tax_table), "Reference Sequences")
+      df_physeq_interm <- cbind(physeq@otu_table,
+                                physeq@tax_table,)
+      colnames(df_physeq_interm) <-
+        c(sample_names(physeq),
+          colnames(physeq@tax_table),
+          "Reference Sequences")
 
       df_physeq_interm <- as.data.frame(df_physeq_interm)
 
       if (write_sam_data) {
-        sam_data <- data.frame(t(data.frame(unclass(physeq@sam_data))))
+        sam_data <- data.frame(t(data.frame(unclass(
+          physeq@sam_data
+        ))))
         colnames(sam_data) <- sample_names(physeq)
         if (sam_data_first) {
           df_physeq <- dplyr::full_join(sam_data, df_physeq_interm)
-          rownames(df_physeq) <- c(rownames(sam_data), rownames(df_physeq_interm))
+          rownames(df_physeq) <-
+            c(rownames(sam_data), rownames(df_physeq_interm))
         } else {
           df_physeq <- dplyr::full_join(df_physeq_interm, sam_data)
-          rownames(df_physeq) <- c(rownames(df_physeq_interm), rownames(sam_data))
+          rownames(df_physeq) <-
+            c(rownames(df_physeq_interm), rownames(sam_data))
         }
       }
       utils::write.table(
@@ -933,8 +986,16 @@ write_pq <- function(physeq,
 #' }
 #' @seealso [MiscMetabar::write_pq()]
 save_pq <- function(physeq, path = NULL, ...) {
-  write_pq(physeq, path = path, rdata = TRUE, one_file = TRUE, ...)
-  write_pq(physeq, path = path, rdata = FALSE, one_file = FALSE, ...)
+  write_pq(physeq,
+           path = path,
+           rdata = TRUE,
+           one_file = TRUE,
+           ...)
+  write_pq(physeq,
+           path = path,
+           rdata = FALSE,
+           one_file = FALSE,
+           ...)
 }
 
 ################################################################################
@@ -960,38 +1021,57 @@ save_pq <- function(physeq, path = NULL, ...) {
 #' read_pq(path = "phyloseq_data")
 #' }
 #'
-read_pq <- function(path = NULL, taxa_are_rows = FALSE, sam_names = NULL, sep_csv = "\t", ...) {
+read_pq <- function(path = NULL,
+                    taxa_are_rows = FALSE,
+                    sam_names = NULL,
+                    sep_csv = "\t",
+                    ...) {
   if (file.exists(paste0(path, "/otu_table.csv"))) {
     if (taxa_are_rows) {
-      otu_table_csv <- as.matrix(utils::read.table(paste0(path, "/otu_table.csv"), sep = sep_csv))
+      otu_table_csv <-
+        as.matrix(utils::read.table(paste0(path, "/otu_table.csv"), sep = sep_csv))
       samp_names <- colnames(otu_table_csv)
       otu_table_csv <- apply(otu_table_csv, 2, as.numeric)
       table_otu <- otu_table(otu_table_csv, taxa_are_rows = TRUE)
       sample_names(table_otu) <- samp_names
       physeq <- phyloseq(table_otu)
     } else {
-      otu_table_csv <- as.matrix(utils::read.table(paste0(path, "/otu_table.csv"), sep = sep_csv))
+      otu_table_csv <-
+        as.matrix(utils::read.table(paste0(path, "/otu_table.csv"), sep = sep_csv))
       samp_names <- rownames(otu_table_csv)
       otu_table_csv <- apply(otu_table_csv, 2, as.numeric)
       rownames(otu_table_csv) <- samp_names
-      physeq <- phyloseq(otu_table(otu_table_csv, taxa_are_rows = FALSE))
+      physeq <-
+        phyloseq(otu_table(otu_table_csv, taxa_are_rows = FALSE))
     }
   }
   if (file.exists(paste0(path, "/refseq.csv"))) {
-    dna <- Biostrings::DNAStringSet(utils::read.table(paste0(path, "/refseq.csv"), sep = sep_csv, row.names = NULL)[, 2])
-    names(dna) <- utils::read.table(paste0(path, "/refseq.csv"), sep = sep_csv, row.names = NULL)[, 1]
+    dna <-
+      Biostrings::DNAStringSet(utils::read.table(
+        paste0(path, "/refseq.csv"),
+        sep = sep_csv,
+        row.names = NULL
+      )[, 2])
+    names(dna) <-
+      utils::read.table(paste0(path, "/refseq.csv"),
+                        sep = sep_csv,
+                        row.names = NULL)[, 1]
     physeq <- phyloseq::merge_phyloseq(physeq, refseq(dna))
   }
   if (file.exists(paste0(path, "/tax_table.csv"))) {
-    tax_table_csv <- utils::read.table(paste0(path, "/tax_table.csv"), sep = sep_csv)
+    tax_table_csv <-
+      utils::read.table(paste0(path, "/tax_table.csv"), sep = sep_csv)
     rownames(tax_table_csv) <- tax_table_csv[, 1]
-    tax_table_csv <- as.matrix(tax_table_csv[, -1])
-    physeq <- phyloseq::merge_phyloseq(physeq, tax_table(tax_table_csv))
+    tax_table_csv <- as.matrix(tax_table_csv[,-1])
+    physeq <-
+      phyloseq::merge_phyloseq(physeq, tax_table(tax_table_csv))
   }
   if (file.exists(paste0(path, "/sam_data.csv"))) {
-    sam_data_csv <- utils::read.table(paste0(path, "/sam_data.csv"), sep = sep_csv)
+    sam_data_csv <-
+      utils::read.table(paste0(path, "/sam_data.csv"), sep = sep_csv)
     rownames(sam_data_csv) <- sam_data_csv[, 1]
-    physeq <- phyloseq::merge_phyloseq(physeq, sample_data(sam_data_csv))
+    physeq <-
+      phyloseq::merge_phyloseq(physeq, sample_data(sam_data_csv))
   }
 
   if (!is.null(physeq@phy_tree)) {
@@ -1112,33 +1192,228 @@ lulu_pq <- function(physeq,
       unlink("match_list.txt")
     }
   }
-  merged <- res_lulu$otu_map[res_lulu$otu_map$curated == "merged", ]
-  merged <- merged[rownames(merged) != merged$parent_id, ]
+  merged <- res_lulu$otu_map[res_lulu$otu_map$curated == "merged",]
+  merged <- merged[rownames(merged) != merged$parent_id,]
 
   test_vector <- vector(mode = "logical")
   for (tax_rank in colnames(physeq@tax_table)) {
-    test <- physeq@tax_table[rownames(merged), tax_rank] == physeq@tax_table[merged$parent_id, tax_rank]
-    test_vector <- c(test_vector, sum(test, na.rm = TRUE) / length(stats::na.exclude(test)))
+    test <-
+      physeq@tax_table[rownames(merged), tax_rank] == physeq@tax_table[merged$parent_id, tax_rank]
+    test_vector <-
+      c(test_vector,
+        sum(test, na.rm = TRUE) / length(stats::na.exclude(test)))
   }
 
   names(test_vector) <- colnames(physeq@tax_table)
 
-  new_physeq <- prune_taxa(taxa_names(physeq) %in% rownames(res_lulu$curated_table), physeq)
-  new_physeq@otu_table <- otu_table(t(res_lulu$curated_table), taxa_are_rows = FALSE)
+  new_physeq <-
+    prune_taxa(taxa_names(physeq) %in% rownames(res_lulu$curated_table),
+               physeq)
+  new_physeq@otu_table <-
+    otu_table(t(res_lulu$curated_table), taxa_are_rows = FALSE)
   sample_names(new_physeq) <- sample_names(physeq)
 
   if (verbose) {
-    message(paste("The number of taxa decrease from ", ntaxa(physeq), " to ", ntaxa(new_physeq), ".", sep = ""))
-    message("See the discrepancy_vector to verify the degree of discrepancy in taxonomy due to lulu re-clustering.")
+    message(paste(
+      "The number of taxa decrease from ",
+      ntaxa(physeq),
+      " to ",
+      ntaxa(new_physeq),
+      ".",
+      sep = ""
+    ))
+    message(
+      "See the discrepancy_vector to verify the degree of discrepancy in taxonomy due to lulu re-clustering."
+    )
   }
-  return(list(
-    "new_physeq" = new_physeq,
-    "discrepancy_vector" = test_vector,
-    "res_lulu" = res_lulu,
-    "merged_ASV" = merged
-  ))
+  return(
+    list(
+      "new_physeq" = new_physeq,
+      "discrepancy_vector" = test_vector,
+      "res_lulu" = res_lulu,
+      "merged_ASV" = merged
+    )
+  )
 }
 ################################################################################
+
+
+################################################################################
+#' MUMU reclustering of class `physeq`
+#'
+#' @description
+#' `r lifecycle::badge("experimental")`
+#'
+#' See https://www.nature.com/articles/s41467-017-01312-x for more information
+#'  on the original method LULU. This is a wrapper of
+#'  [mumu](https://github.com/frederic-mahe/mumu) a C++ re-implementation
+#'  of LULU by Frédéric Mahé
+#'
+#' @inheritParams clean_pq
+#' @param nproc (default 1)
+#'   Set to number of cpus/processors to use for the clustering
+#' @param id (default: 0.84) id for --usearch_global.
+#' @param vsearchpath path to vsearch.
+#' @param vsearchpath path to mumu. See [mumu](https://github.com/frederic-mahe/mumu)
+#'   for installation instruction
+#' @param verbose (logical) if true, print some additional messages.
+#' @param clean_pq (logical) if true, empty samples and empty ASV are discarded
+#'   before clustering.
+#' @param  keep_temporary_files (logical, default: FALSE) Do we keep temporary files
+#' @return a list of for object
+#' - "new_physeq": The new phyloseq object (class physeq)
+#' - "mumu_results": The log file of the mumu software. Run `man mumu` into
+#'   bash to obtain details about columns' signification.
+#'
+#' @export
+#' @examples
+#' \dontrun{
+#' data(data_fungi_sp_known)
+#' mumu_pq(data_fungi_sp_known)
+#' }
+#' @author Frédéric Mahé \url{"https://github.com/frederic-mahe/"}
+#'   & Adrien Taudière \email{adrien.taudiere@@zaclys.net}
+#' @references
+#' - MUMU: \url{https://github.com/frederic-mahe/mumu}
+#' - VSEARCH can be downloaded from
+#'  \url{https://github.com/torognes/vsearch}.
+#' @details
+#' This function is mainly a wrapper of the work of others.
+#'   Please cite [mumu](https://github.com/frederic-mahe/mumu/blob/main/CITATION.cff) and
+#'   [lulu](https://www.nature.com/articles/s41467-017-01312-x) if you use this function
+#'   for your work.
+#'
+mumu_pq <- function(physeq,
+                    nproc = 1,
+                    id = 0.84,
+                    vsearchpath = "vsearch",
+                    mumupath = "mumu",
+                    verbose = FALSE,
+                    clean_pq = TRUE,
+                    keep_temporary_files = FALSE) {
+  verify_pq(physeq)
+  if (is.null(physeq@refseq)) {
+    stop("The phyloseq object do not contain a @refseq slot")
+  }
+  if (clean_pq) {
+    physeq <- clean_pq(physeq)
+  }
+
+  message("Start Vsearch usearch_global")
+  dna <- Biostrings::DNAStringSet(physeq@refseq)
+  Biostrings::writeXStringSet(dna, "temp.fasta")
+  system2(
+    vsearchpath,
+    paste(
+      " --usearch_global temp.fasta --db temp.fasta --self --iddef 1",
+      " -userfields query+target+id --maxaccepts 0 --query_cov 0.9 --maxhits 10",
+      " -id ",
+      id,
+      "  --userout match_list.txt"
+    ),
+    stdout = TRUE,
+    stderr = TRUE
+  )
+  otu_tab <-
+    data.frame(unclass(clean_pq(physeq, force_taxa_as_rows = TRUE)@otu_table))
+  otu_tab <- cbind("ASV" = rownames(otu_tab), otu_tab)
+  write.table(
+    otu_tab,
+    "otu_table.csv",
+    sep = "\t",
+    row.names = FALSE,
+    quote = FALSE
+  )
+
+  message("Mumu algorithm")
+  system2(
+    mumupath,
+    "--otu_table otu_table.csv --match_list match_list.txt --log log.txt --new_otu_table new_OTU.tablemumu"
+  )
+
+  res_mumu <- read.delim("new_OTU.tablemumu")
+  new_otu_tab <- otu_table(res_mumu[, -1], taxa_are_rows = TRUE)
+  taxa_names(new_otu_tab) <- res_mumu[, 1]
+
+  new_physeq <-
+    prune_taxa(taxa_names(physeq) %in% taxa_names(new_otu_tab), physeq)
+  new_physeq@otu_table <-
+    otu_table(t(new_otu_tab), taxa_are_rows = FALSE)
+  if (nsamples(new_physeq) != nsamples(physeq)) {
+    stop(
+      "There is a different number of samples before and after mumu algorithm.
+         This may be due to empty samples. You may try to rerun mumu_pq()
+         using clean_pq = TRUE."
+    )
+  }
+  sample_names(new_physeq) <- sample_names(physeq)
+
+  if (verbose) {
+    message(paste(
+      "The number of taxa decrease from ",
+      ntaxa(physeq),
+      " to ",
+      ntaxa(new_physeq),
+      ".",
+      sep = ""
+    ))
+    message(
+      "See the log slot to verify the degree of discrepancy in taxonomy due to mumu re-clustering."
+    )
+  }
+
+  result_mumu <- read.delim("log.txt")
+  colnames(result_mumu) <-
+    c(
+      "Query_ASV",
+      "Potential parent",
+      "Similarity_percent",
+      "Ab_query",
+      "Ab_parent",
+      "Overlap_ab_query",
+      "Overlap_ab_parent",
+      "Incidence_query",
+      "Incidence_parent",
+      "Smallest_ab_ratio",
+      "Sum_ab_ratio",
+      "Average_ratio",
+      "Average_non_null_ratio",
+      "Relative_cooccurence_value",
+      "Status"
+    )
+
+  if (!keep_temporary_files) {
+    if (file.exists("temp.fasta")) {
+      unlink("temp.fasta")
+    }
+    if (file.exists("cluster.fasta")) {
+      unlink("cluster.fasta")
+    }
+    if (file.exists("temp.uc")) {
+      unlink("temp.uc")
+    }
+  }
+  if (file.exists("log.txt")) {
+    unlink("temp.uc")
+  }
+  if (file.exists("match_list.txt")) {
+    unlink("match_list.txt")
+  }
+  if (file.exists("otu_table.csv")) {
+    unlink("otu_table.csv")
+  }
+  if (file.exists("new_OTU.tablemumu")) {
+    unlink("new_OTU.tablemumu")
+  }
+}
+
+return(list("new_physeq" = new_physeq,
+            "mumu_results" = result_mumu))
+}
+################################################################################
+
+
+
 
 ################################################################################
 #' Verify the validity of a phyloseq object
@@ -1154,7 +1429,8 @@ lulu_pq <- function(physeq,
 #' @export
 #'
 verify_pq <- function(physeq) {
-  if (!methods::validObject(physeq) || !inherits(physeq, "phyloseq")) {
+  if (!methods::validObject(physeq) ||
+      !inherits(physeq, "phyloseq")) {
     stop("The physeq argument is not a valid phyloseq object.")
   }
 }
@@ -1202,7 +1478,7 @@ subset_samples_pq <- function(physeq, condition) {
     return(physeq)
   } else {
     old_DF <- as(sample_data(physeq), "data.frame")
-    new_DF <- old_DF[condition, ]
+    new_DF <- old_DF[condition,]
     if (inherits(physeq, "sample_data")) {
       return(sample_data(new_DF))
     } else {
@@ -1242,7 +1518,10 @@ subset_samples_pq <- function(physeq, condition) {
 #' @return a new phyloseq object
 #' @export
 #'
-subset_taxa_pq <- function(physeq, condition, verbose = TRUE, clean_pq = TRUE) {
+subset_taxa_pq <- function(physeq,
+                           condition,
+                           verbose = TRUE,
+                           clean_pq = TRUE) {
   if (inherits(condition, "taxonomyTable")) {
     condition_temp <- as.vector(condition)
     names(condition_temp) <- rownames(condition)
@@ -1259,7 +1538,8 @@ subset_taxa_pq <- function(physeq, condition, verbose = TRUE, clean_pq = TRUE) {
   new_physeq <- physeq
 
   if (!taxa_are_rows(new_physeq)) {
-    new_physeq@otu_table <- otu_table(t(new_physeq@otu_table), taxa_are_rows = TRUE)
+    new_physeq@otu_table <-
+      otu_table(t(new_physeq@otu_table), taxa_are_rows = TRUE)
     taxa_are_rows(new_physeq) <- TRUE
   }
 
@@ -1267,7 +1547,7 @@ subset_taxa_pq <- function(physeq, condition, verbose = TRUE, clean_pq = TRUE) {
   cond[is.na(cond)] <- FALSE
 
   old_MA <- as(otu_table(new_physeq), "matrix")
-  new_MA <- old_MA[cond, ]
+  new_MA <- old_MA[cond,]
 
   if (!is.matrix(new_MA)) {
     new_MA <- as.matrix(new_MA)
@@ -1284,9 +1564,16 @@ subset_taxa_pq <- function(physeq, condition, verbose = TRUE, clean_pq = TRUE) {
   }
 
   if (verbose) {
-    message(paste("Number of non-matching ASV", sum(is.na(match(taxa_names(physeq), names(condition))))))
-    message(paste("Number of matching ASV", sum(!is.na(match(taxa_names(physeq), names(condition))))))
-    message(paste("Number of filtered-out ASV", ntaxa(physeq) - ntaxa(new_physeq)))
+    message(paste("Number of non-matching ASV", sum(is.na(
+      match(taxa_names(physeq), names(condition))
+    ))))
+    message(paste("Number of matching ASV", sum(!is.na(
+      match(taxa_names(physeq), names(condition))
+    ))))
+    message(paste(
+      "Number of filtered-out ASV",
+      ntaxa(physeq) - ntaxa(new_physeq)
+    ))
     message(paste("Number of kept ASV", ntaxa(new_physeq)))
     message(paste("Number of kept samples", nsamples(new_physeq)))
   }
@@ -1320,12 +1607,36 @@ subset_taxa_pq <- function(physeq, condition, verbose = TRUE, clean_pq = TRUE) {
 #' A8_005
 select_one_sample <- function(physeq, sam_name, silent = FALSE) {
   if (sum(sample_names(physeq) %in% sam_name) == 0) {
-    stop(paste0("The sample ", sam_name, " is not present in the names of samples of your phyloseq physeq object. You may use the sample_names() function."))
+    stop(
+      paste0(
+        "The sample ",
+        sam_name,
+        " is not present in the names of samples of your phyloseq physeq object. You may use the sample_names() function."
+      )
+    )
   }
-  cl_sam <- clean_pq(subset_samples_pq(physeq, sample_names(physeq) == sam_name), silent = TRUE)
+  cl_sam <-
+    clean_pq(subset_samples_pq(physeq, sample_names(physeq) == sam_name),
+             silent = TRUE)
 
   if (!silent) {
-    message(paste0("You select 1 of ", nsamples(physeq), " samples and conserved ", ntaxa(cl_sam), " out of ", ntaxa(physeq), " taxa represented by ", sum(cl_sam@otu_table), " sequences (out of ", sum(physeq@otu_table), " sequences [", perc(sum(cl_sam@otu_table), sum(physeq@otu_table)), "%])"))
+    message(
+      paste0(
+        "You select 1 of ",
+        nsamples(physeq),
+        " samples and conserved ",
+        ntaxa(cl_sam),
+        " out of ",
+        ntaxa(physeq),
+        " taxa represented by ",
+        sum(cl_sam@otu_table),
+        " sequences (out of ",
+        sum(physeq@otu_table),
+        " sequences [",
+        perc(sum(cl_sam@otu_table), sum(physeq@otu_table)),
+        "%])"
+      )
+    )
   }
 
   return(cl_sam)
@@ -1362,8 +1673,10 @@ add_new_taxonomy_pq <- function(physeq, ref_fasta, suffix = NULL, ...) {
   if (is.null(suffix)) {
     suffix <- basename(ref_fasta)
   }
-  tax_tab <- dada2::assignTaxonomy(physeq@refseq, refFasta = ref_fasta, ...)
-  colnames(tax_tab) <- make.unique(paste0(colnames(tax_tab), "_", suffix))
+  tax_tab <-
+    dada2::assignTaxonomy(physeq@refseq, refFasta = ref_fasta, ...)
+  colnames(tax_tab) <-
+    make.unique(paste0(colnames(tax_tab), "_", suffix))
   new_tax_tab <- tax_table(cbind(physeq@tax_table, tax_tab))
   new_physeq <- physeq
   tax_table(new_physeq) <- new_tax_tab
@@ -1378,35 +1691,39 @@ add_new_taxonomy_pq <- function(physeq, ref_fasta, suffix = NULL, ...) {
 #' @description
 #' `r lifecycle::badge("experimental")`
 #'
-#' A wrapper for the [tbl_summary::tbl_summary(...)] function in the case of `physeq`
+#' A wrapper for the [gtsummary::tbl_summary()] function in the case of `physeq`
 #'   object.
-#' 
+#'
 #' @inheritParams clean_pq
-#' @param remove_col_unique_value (logical, default TRUE) Do we remove informative 
-#'   columns (categorical column with one value per samples), e.g. samples names ?  
-#' @param ... Others arguments pass on to [tbl_summary::tbl_summary()].
+#' @param remove_col_unique_value (logical, default TRUE) Do we remove informative
+#'   columns (categorical column with one value per samples), e.g. samples names ?
+#' @param ... Others arguments pass on to [gtsummary::tbl_summary()].
 #' @return a physeq object with a larger slot tax_table
 #'
 #' @export
 #' @author Adrien Taudière
 #' @examples
 #' tbl_sum_samdata(data_fungi)
-#' 
-#' tbl_sum_samdata(data_fungi, include = c("Time", "Height"), 
-#'                 type= list(Time ~ "continuous2", Height ~ "categorical"), 
+#'
+#' tbl_sum_samdata(data_fungi, include = c("Time", "Height"),
+#'                 type= list(Time ~ "continuous2", Height ~ "categorical"),
 #'                 statistic = list(Time ~ c("{median} ({p25}, {p75})", "{min}, {max}")))
-#' 
+#'
 #' tbl_sum_samdata (enterotype)
 #' tbl_sum_samdata (enterotype, include =  !contains("SampleId"))
+#' @details
+#' This function is mainly a wrapper of the work of others.
+#'   Please make a reference to `gtsummary::tbl_summary()` if you
+#'   use this function.
 
 tbl_sum_samdata <- function(physeq, remove_col_unique_value = TRUE, ...) {
   tbl <- tibble(data.frame(physeq@sam_data))
   if (remove_col_unique_value) {
-    tbl <- tbl[, !apply(tbl, 2, function(x) {
+    tbl <- tbl[,!apply(tbl, 2, function(x) {
       length(unique(x)) == nrow(tbl) && is.character(x)
-      })]
+    })]
   }
-  tbl_sum <- tbl %>% tbl_summary::tbl_summary(...)
+  tbl_sum <- tbl %>% gtsummary::tbl_summary(...)
   return(tbl_sum)
 }
 ################################################################################